+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Influxdb is now deployed from a fork of the stable/influxdb chart, instead of using Kustomize to modify the original.
 - Pipeline only tests Harbor if `ENABLE_HARBOR` is set to true.
 - Refactorization of configuration repositories.
+- Provided more meaningful context to kubeconfig names
 
 ### Removed